+++ administrator/templates/atum/login.php
@@ -10,6 +10,8 @@
 defined('_JEXEC') or die;
 
 use Joomla\CMS\Factory;
+use Joomla\CMS\Uri\Uri;
+use Joomla\CMS\Language\Text;
 use Joomla\CMS\HTML\HTMLHelper;
 
 /** @var JDocumentHtml $this */
@@ -54,79 +56,53 @@ $this->setMetaData('theme-color', '#1c3d5c');
 <head>
 	<jdoc:include type="metas" />
 	<jdoc:include type="styles" />
-	<style>
-		.login-initial {
-			display: none;
-		}
-		<?php // Check if debug is on ?>
-		<?php if ($app->get('debug_lang', 1) || $app->get('debug', 1)) : ?>
-		.view-login .container {
-			position: static;
-			margin-top: 20px;
-			margin-left: auto;
-			margin-right: auto;
-		}
-		.view-login .navbar-fixed-bottom {
-			display: none;
-		}
-		<?php endif; ?>
-	</style>
 </head>
 
 <body class="site <?php echo $option . ' view-' . $view . ' layout-' . $layout . ' task-' . $task . ' itemid-' . $itemid . ' '; ?>">
 	<?php // Container ?>
-	<div class="container">
-		<div class="login-logo">
-			<img class="card-img-top" src="<?php echo $this->baseurl; ?>/templates/<?php echo $this->template; ?>/images/logo.svg" alt="<?php echo $sitename; ?>">
-		</div>
-		<div id="content">
-			<noscript>
-				<div class="alert alert-danger" role="alert">
-					<?php echo JText::_('JGLOBAL_WARNJAVASCRIPT'); ?>
-				</div>
-			</noscript>
-			<?php // Begin Content ?>
-			<div id="element-box" class="login card">
-				<div class="card-body">
-					<h1 class="text-center mt-1 mb-4"><?php echo JText::_('MOD_LOGIN_LOGIN_TITLE'); ?></h1>
-					<jdoc:include type="message" />
-					<jdoc:include type="component" />
-				</div>
+	<div class="d-flex justify-content-center align-items-center h-100">
+		<div class="login">
+			<div class="login-logo">
+				<img class="card-img-top" src="<?php echo $this->baseurl; ?>/templates/<?php echo $this->template; ?>/images/logo.svg" alt="<?php echo $sitename; ?>">
 			</div>
-			<?php // End Content ?>
+			<div id="content">
+				<noscript>
+					<div class="alert alert-danger" role="alert">
+						<?php echo Text::_('JGLOBAL_WARNJAVASCRIPT'); ?>
+					</div>
+				</noscript>
+				<?php // Begin Content ?>
+				<div id="element-box" class="card">
+					<div class="card-body">
+						<h1 class="text-center mt-1 mb-4"><?php echo Text::_('MOD_LOGIN_LOGIN_TITLE'); ?></h1>
+						<jdoc:include type="message" />
+						<jdoc:include type="component" />
+					</div>
+				</div>
+				<?php // End Content ?>
+			</div>
 		</div>
 	</div>
 
-	<nav class="navbar fixed-bottom hidden-xs-down">
-		<ul class="nav nav-fill">
-			<li class="nav-item">
-				<a href="<?php echo JUri::root(); ?>" target="_blank" class="float-left"><span class="fa fa-external-link" aria-hidden="true"></span> <?php echo JText::_('COM_LOGIN_RETURN_TO_SITE_HOME_PAGE'); ?></a>
-			</li>
-			<li class="nav-item">
-				<a class="login-joomla hasTooltip" href="https://www.joomla.org" target="_blank" title="<?php echo JHtml::tooltipText('TPL_ATUM_ISFREESOFTWARE'); ?>">
-					<span class="fa fa-joomla" aria-hidden="true"></span>
-					<span class="sr-only"><?php echo JText::_('TPL_ATUM_GOTO_JOOMLA_HOME_PAGE'); ?></span>
+	<div class="fixed-bottom px-3 mb-2 d-none d-md-block">
+		<div class="row nav align-items-center">
+			<div class="col">
+				<a href="<?php echo Uri::root(); ?>" target="_blank"><span class="fa fa-external-link mr-1" aria-hidden="true"></span><?php echo Text::_('COM_LOGIN_RETURN_TO_SITE_HOME_PAGE'); ?></a>
+			</div>
+			<div class="col text-center">
+				<a href="https://www.joomla.org" target="_blank" title="<?php echo Text::_('TPL_ATUM_ISFREESOFTWARE'); ?>">
+					<span class="fa fa-2x fa-joomla" aria-hidden="true"></span>
+					<span class="sr-only"><?php echo Text::_('TPL_ATUM_GOTO_JOOMLA_HOME_PAGE'); ?></span>
 				</a>
-			</li>
-			<li class="nav-item">
-				<span class="text-white float-right">&nbsp;&copy; <?php echo date('Y'); ?> <?php echo $sitename; ?></span>
-			</li>
-		</ul>
-	</nav>
+			</div>
+			<div class="col text-right">
+				<span class="text-white">&nbsp;&copy; <?php echo date('Y'); ?> <?php echo $sitename; ?></span>
+			</div>
+		</div>
+	</div>
 
 	<jdoc:include type="modules" name="debug" style="none" />
 
 	<jdoc:include type="scripts" />
-	<script>
-		(function() {
-			var formTmp = document.querySelector('.login-initial');
-			if (formTmp) {
-				formTmp.style.display = 'block';
-				if (!document.querySelector('joomla-alert')) {
-					document.getElementById('mod-login-username').focus();
-				}
-			}
-		})();
-	</script>
 </body>
 </html>
