+++ administrator/templates/atum/scss/blocks/_global.scss
@@ -2,6 +2,7 @@
 
 html,
 body {
+  height: 100%;
   font-size: $font-size-root;
 }
